changed name from onCreate ot createCallback

+++ src/components/footer/Footer.tsx
@@ -11,6 +11,7 @@ type Props = {
   setIsError: (el: boolean) => void;
   ErrorMessages: { None: string; Delete: string };
   setErrorMessage: (msg: string) => void;
+  USER_ID: number;
 };
 
 export const Footer: React.FC<Props> = ({
@@ -18,6 +19,7 @@ export const Footer: React.FC<Props> = ({
   setTodos,
   statusFilter,
   setStatusFilter,
+  USER_ID,
 }) => {
   const setStatus = (
     val: string,
@@ -31,7 +33,7 @@ export const Footer: React.FC<Props> = ({
     const completedTodos = todos?.filter(todo => todo.completed) || [];
 
     const deletePromises = completedTodos.map(todo =>
-      client.delete(`/todos/${todo.id}`),
+      client.delete(`/todos/${todo.id}?userId=${USER_ID}`),
     );
 
     await Promise.all(deletePromises);
